requirements.txt
openenv-core>=0.2.2
openai>=1.0.0
fastapi
uvicorn
pydantic
numpy
gradio